+++ demo-formatter/go/cucumber_demo_formatter_test.go
@@ -12,14 +12,14 @@ func TestAllResultTypes(t *testing.T) {
 	stdin := &bytes.Buffer{}
 	writer := fio.NewNdjsonWriter(stdin)
 
-	var statuses = []messages.TestResult_Status{
-		messages.TestResult_UNKNOWN,
-		messages.TestResult_PASSED,
-		messages.TestResult_SKIPPED,
-		messages.TestResult_PENDING,
-		messages.TestResult_UNDEFINED,
-		messages.TestResult_AMBIGUOUS,
-		messages.TestResult_FAILED,
+	var statuses = []messages.TestStepResult_Status{
+		messages.TestStepResult_UNKNOWN,
+		messages.TestStepResult_PASSED,
+		messages.TestStepResult_SKIPPED,
+		messages.TestStepResult_PENDING,
+		messages.TestStepResult_UNDEFINED,
+		messages.TestStepResult_AMBIGUOUS,
+		messages.TestStepResult_FAILED,
 	}
 	for _, status := range statuses {
 		err := writer.WriteMsg(newTestStepFinished(status))
@@ -42,11 +42,11 @@ func TestAllResultTypes(t *testing.T) {
 		stdout.String())
 }
 
-func newTestStepFinished(status messages.TestResult_Status) *messages.Envelope {
+func newTestStepFinished(status messages.TestStepResult_Status) *messages.Envelope {
 	return &messages.Envelope{
 		Message: &messages.Envelope_TestStepFinished{
 			TestStepFinished: &messages.TestStepFinished{
-				TestResult: &messages.TestResult{
+				TestStepResult: &messages.TestStepResult{
 					Status: status,
 				},
 			},